[Infra] Remove HeartbeatLoggingTestUtils.podspec
