+++ FirebaseCore.podspec
@@ -58,26 +58,21 @@ Firebase Core includes FIRApp and FIROptions which provide central configuration
     'OTHER_CFLAGS' => '-fno-autolink'
   }
 
-  # Using environment variable because of the dependency on the unpublished
-  # HeartbeatLoggingTestUtils.
-  if ENV['POD_LIB_LINT_ONLY'] && ENV['POD_LIB_LINT_ONLY'] == '1' then
-    s.test_spec 'unit' do |unit_tests|
-      unit_tests.scheme = { :code_coverage => true }
-      unit_tests.platforms = {
-        :ios => ios_deployment_target,
-        :osx => '10.15',
-        :tvos => tvos_deployment_target
-      }
-      unit_tests.source_files = [
-        'FirebaseCore/Tests/Unit/**/*.[mh]',
-        'SharedTestUtilities/FIROptionsMock.[mh]',
-      ]
+  s.test_spec 'unit' do |unit_tests|
+    unit_tests.scheme = { :code_coverage => true }
+    unit_tests.platforms = {
+      :ios => ios_deployment_target,
+      :osx => '10.15',
+      :tvos => tvos_deployment_target
+    }
+    unit_tests.source_files = [
+      'FirebaseCore/Tests/Unit/**/*.[mh]',
+      'SharedTestUtilities/FIROptionsMock.[mh]',
+    ]
 
-      unit_tests.requires_app_host = true
-      unit_tests.dependency 'OCMock'
-      unit_tests.dependency 'HeartbeatLoggingTestUtils'
-      unit_tests.resources = 'FirebaseCore/Tests/Unit/Resources/GoogleService-Info.plist'
-    end
+    unit_tests.requires_app_host = true
+    unit_tests.dependency 'OCMock'
+    unit_tests.resources = 'FirebaseCore/Tests/Unit/Resources/GoogleService-Info.plist'
   end
 
   s.test_spec 'swift-unit' do |swift_unit_tests|
