+++ FirebaseInstallations.podspec
@@ -54,43 +54,37 @@ Pod::Spec.new do |s|
     'HEADER_SEARCH_PATHS' => '"${PODS_TARGET_SRCROOT}"'
   }
 
-  # Using environment variable because of the dependency on the unpublished
-  # HeartbeatLoggingTestUtils.
-  if ENV['POD_LIB_LINT_ONLY'] && ENV['POD_LIB_LINT_ONLY'] == '1' then
-    s.test_spec 'unit' do |unit_tests|
-      unit_tests.scheme = { :code_coverage => true }
-      unit_tests.platforms = {
-        :ios => ios_deployment_target,
-        :osx => '10.15',
-        :tvos => tvos_deployment_target
+  s.test_spec 'unit' do |unit_tests|
+    unit_tests.scheme = { :code_coverage => true }
+    unit_tests.platforms = {
+      :ios => ios_deployment_target,
+      :osx => '10.15',
+      :tvos => tvos_deployment_target
+    }
+    unit_tests.source_files = base_dir + 'Tests/Unit/*.[mh]',
+                              base_dir + 'Tests/Utils/*.[mh]',
+                              base_dir + 'Tests/Unit/Swift/*'
+    unit_tests.resources = base_dir + 'Tests/Fixture/**/*'
+    unit_tests.requires_app_host = true
+    unit_tests.dependency 'OCMock'
+
+    if ENV['FIS_IID_MIGRATION_TESTING'] && ENV['FIS_IID_MIGRATION_TESTING'] == '1' then
+      unit_tests.source_files += base_dir + 'Tests/Unit/IIDStoreTests/*.[mh]'
+    end
+  end
+
+  s.test_spec 'integration' do |int_tests|
+    int_tests.scheme = { :code_coverage => true }
+    int_tests.platforms = {:ios => '10.0', :osx => '10.15', :tvos => '11.0'}
+    int_tests.source_files = base_dir + 'Tests/Integration/**/*.[mh]'
+    int_tests.resources = base_dir + 'Tests/Resources/**/*'
+    if ENV['FIS_INTEGRATION_TESTS_REQUIRED'] && ENV['FIS_INTEGRATION_TESTS_REQUIRED'] == '1' then
+      int_tests.pod_target_xcconfig = {
+      'GCC_PREPROCESSOR_DEFINITIONS' =>
+        'FIR_INSTALLATIONS_INTEGRATION_TESTS_REQUIRED=1'
       }
-      unit_tests.source_files = base_dir + 'Tests/Unit/*.[mh]',
-                                base_dir + 'Tests/Utils/*.[mh]',
-                                base_dir + 'Tests/Unit/Swift/*'
-      unit_tests.resources = base_dir + 'Tests/Fixture/**/*'
-      unit_tests.requires_app_host = true
-      unit_tests.dependency 'OCMock'
-      unit_tests.dependency 'HeartbeatLoggingTestUtils'
-
-      if ENV['FIS_IID_MIGRATION_TESTING'] && ENV['FIS_IID_MIGRATION_TESTING'] == '1' then
-        unit_tests.source_files += base_dir + 'Tests/Unit/IIDStoreTests/*.[mh]'
-      end
     end
-
-    s.test_spec 'integration' do |int_tests|
-      int_tests.scheme = { :code_coverage => true }
-      int_tests.platforms = {:ios => '10.0', :osx => '10.15', :tvos => '11.0'}
-      int_tests.source_files = base_dir + 'Tests/Integration/**/*.[mh]'
-      int_tests.resources = base_dir + 'Tests/Resources/**/*'
-      if ENV['FIS_INTEGRATION_TESTS_REQUIRED'] && ENV['FIS_INTEGRATION_TESTS_REQUIRED'] == '1' then
-        int_tests.pod_target_xcconfig = {
-        'GCC_PREPROCESSOR_DEFINITIONS' =>
-          'FIR_INSTALLATIONS_INTEGRATION_TESTS_REQUIRED=1'
-        }
-      end
-      int_tests.requires_app_host = true
-      int_tests.dependency 'OCMock'
-      int_tests.dependency 'HeartbeatLoggingTestUtils'
-    end
+    int_tests.requires_app_host = true
+    int_tests.dependency 'OCMock'
   end
 end
